chore: update import statements for handlers to reflect HttpHandler move
